pyproject.toml
[project]
name = "rag-cli-tool"
version = "0.1.0"
description = "User-friendly document RAG CLI with LightRAG-style workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
rag = [
  "lightrag-hku>=1.3.0",
]
doc = [
  "docling>=2.0.0",
]

[project.scripts]
rag-cli = "rag_cli.cli.main:app"

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_cli"]

[tool.uv]
package = true

[tool.uv.sources]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

